backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
graphene-django==3.1.5
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
django-extensions==3.2.3
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2 